comment
Comments are very useful in any language to tell a user what is the task of any function or operation. The comments are neglected by the compiler, so whatever you write in the comments won't have any effect on the working of a code. In C/C++, comments can be written as mentioned below:

/*This is a comment*/
//This is also a comment

Example:

Input:
a = 5
b = 6
c = 15
Output:
5
15
Explanation:
Value of b is not printed.
 

Your Task:
Your task is to complete a function that print a, b and c. Comment the code that outputs b, so only a and c gets printed.

Constraints:
1 <= a <= 106
1 <=b <= 106
1 <= c<= 106